+++ braindecode/models/__init__.py
@@ -1,12 +1,13 @@
 """
 Some predefined network architectures for EEG decoding.
 """
-
+from .eegitnet import EEGITNet
 from .deep4 import Deep4Net
 from .eegnet import EEGNetv4, EEGNetv1
 from .hybrid import HybridNet
 from .shallow_fbcsp import ShallowFBCSPNet
 from .eegresnet import EEGResNet
+from .eeginception import EEGInception
 from .tcn import TCN
 from .sleep_stager_chambon_2018 import SleepStagerChambon2018
 from .sleep_stager_blanco_2020 import SleepStagerBlanco2020
@@ -15,3 +16,4 @@ from .tidnet import TIDNet
 from .usleep import USleep
 from .util import get_output_shape, to_dense_prediction_model
 from .modules import TimeDistributed
+