+++ test/unit_tests/models/test_models.py
@@ -12,7 +12,9 @@ import pytest
 
 from braindecode.models import (
     Deep4Net, EEGNetv4, EEGNetv1, HybridNet, ShallowFBCSPNet, EEGResNet, TCN,
-    SleepStagerChambon2018, SleepStagerBlanco2020, SleepStagerEldele2021, USleep, TIDNet)
+    SleepStagerChambon2018, SleepStagerBlanco2020, SleepStagerEldele2021, USleep,
+    EEGITNet, EEGInception, TIDNet)
+
 from braindecode.util import set_random_seeds
 
 
@@ -110,8 +112,47 @@ def test_tcn(input_sizes):
     check_forward_pass(model, input_sizes, only_check_until_dim=2)
 
 
-@pytest.mark.parametrize('n_channels,sfreq,n_classes,input_size_s',
-                         [(20, 128, 5, 30), (10, 256, 4, 20), (1, 64, 2, 30)])
+def test_eegitnet(input_sizes):
+    model = EEGITNet(
+        n_classes=input_sizes['n_classes'],
+        in_channels=input_sizes['n_channels'],
+        input_window_samples=input_sizes['n_in_times'])
+
+    check_forward_pass(model, input_sizes,)
+
+
+def test_eeginception(input_sizes):
+    model = EEGInception(
+        n_classes=input_sizes['n_classes'],
+        in_channels=input_sizes['n_channels'],
+        input_window_samples=input_sizes['n_in_times'])
+
+    check_forward_pass(model, input_sizes,)
+
+
+def test_eeginception_n_params():
+    """Make sure the number of parameters is the same as in the paper when
+    using the same architecture hyperparameters.
+    """
+    model = EEGInception(
+        in_channels=8,
+        n_classes=2,
+        input_window_samples=128,  # input_time
+        sfreq=128,
+        drop_prob=0.5,
+        n_filters=8,
+        scales_samples_s=(0.5, 0.25, 0.125),
+        activation=torch.nn.ELU()
+    )
+
+    n_params = sum(p.numel() for p in model.parameters() if p.requires_grad)
+    assert n_params == 14926  # From paper's TABLE IV EEG-Inception Architecture Details
+
+
+@pytest.mark.parametrize(
+    "n_channels,sfreq,n_classes,input_size_s",
+    [(20, 128, 5, 30), (10, 256, 4, 20), (1, 64, 2, 30)],
+)
 def test_sleep_stager(n_channels, sfreq, n_classes, input_size_s):
     rng = np.random.RandomState(42)
     time_conv_size_s = 0.5
@@ -132,12 +173,15 @@ def test_sleep_stager(n_channels, sfreq, n_classes, input_size_s):
     y_pred2 = model(X.unsqueeze(1))  # 4D inputs
     assert y_pred1.shape == (n_examples, n_classes)
     assert y_pred2.shape == (n_examples, n_classes)
-    np.testing.assert_allclose(y_pred1.detach().cpu().numpy(),
-                               y_pred2.detach().cpu().numpy())
+    np.testing.assert_allclose(
+        y_pred1.detach().cpu().numpy(), y_pred2.detach().cpu().numpy()
+    )
 
 
-@pytest.mark.parametrize('in_chans,sfreq,n_classes,input_size_s',
-                         [(20, 128, 5, 30), (10, 100, 4, 20), (1, 64, 2, 30)])
+@pytest.mark.parametrize(
+    "in_chans,sfreq,n_classes,input_size_s",
+    [(20, 128, 5, 30), (10, 100, 4, 20), (1, 64, 2, 30)],
+)
 def test_usleep(in_chans, sfreq, n_classes, input_size_s):
     rng = np.random.RandomState(42)
     n_examples = 10
@@ -279,3 +323,24 @@ def test_blanco_2020_feats():
 
     out = model(X)
     assert out.shape == (n_examples, model.len_last_layer)
+
+
+def test_eegitnet_shape():
+    n_channels = 2
+    sfreq = 50
+    input_size_s = 30
+    n_classes = 3
+    n_examples = 10
+    model = EEGITNet(
+        n_classes=n_classes,
+        in_channels=n_channels,
+        input_window_samples=int(sfreq * input_size_s),
+    )
+    model.eval()
+
+    rng = np.random.RandomState(42)
+    X = rng.randn(n_examples, n_channels, int(sfreq * input_size_s))
+    X = torch.from_numpy(X.astype(np.float32))
+
+    out = model(X)
+    assert out.shape == (n_examples, n_classes)